# docs/openapi.yaml
openapi: 3.0.1
info:
  title: ShopAPI
  description: "API REST de gestion de pedidos e inventario, con autenticacion JWT,\
    \ control de stock y maquina de estados de pedidos."
  contact:
    name: Kike
    url: https://github.com/enriquegar95/shopapi
  license:
    name: MIT
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
tags:
- name: Productos
  description: Gestion del catalogo de productos
paths:
  /api/v1/usuarios/{id}:
    get:
      tags:
      - usuario-controller
      operationId: obtenerPorId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UsuarioResponseDTO"
    put:
      tags:
      - usuario-controller
      operationId: actualizar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UsuarioResponseDTO"
    delete:
      tags:
      - usuario-controller
      operationId: eliminar
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/productos/{id}:
    get:
      tags:
      - Productos
      summary: Obtener un producto por id
      operationId: obtenerPorId_1
      parameters:
      - name: id
        in: path
        description: Id del producto
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Producto no encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductoResponseDTO"
        "200":
          description: Producto encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductoResponseDTO"
    put:
      tags:
      - Productos
      summary: Actualizar un producto existente
      operationId: actualizar_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductoRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductoResponseDTO"
    delete:
      tags:
      - Productos
      summary: Eliminar un producto
      description: Solo el rol ADMIN puede eliminar productos
      operationId: eliminar_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/usuarios:
    get:
      tags:
      - usuario-controller
      operationId: listar
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageUsuarioResponseDTO"
    post:
      tags:
      - usuario-controller
      operationId: crear
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UsuarioResponseDTO"
  /api/v1/productos:
    get:
      tags:
      - Productos
      summary: Listar productos
      description: "Listado paginado de todo el catalogo. Requiere estar autenticado,\
        \ cualquier rol"
      operationId: listar_1
      parameters:
      - name: pageable
        in: query
        description: "Paginacion y orden, ej: page=0&size=10&sort=nombre,asc"
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageProductoResponseDTO"
    post:
      tags:
      - Productos
      summary: Crear un producto
      description: Solo los roles ADMIN o VENDEDOR pueden crear productos nuevos en
        el catalogo
      operationId: crear_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductoRequestDTO"
        required: true
      responses:
        "400":
          description: Datos invalidos (ver el detalle de cada campo)
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductoResponseDTO"
        "401":
          description: No autenticado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductoResponseDTO"
        "403":
          description: Autenticado pero sin permisos suficientes
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductoResponseDTO"
        "201":
          description: Producto creado correctamente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductoResponseDTO"
        "404":
          description: La categoria indicada no existe
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductoResponseDTO"
  /api/v1/pedidos:
    get:
      tags:
      - pedido-controller
      operationId: listar_2
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagePedidoResponseDTO"
    post:
      tags:
      - pedido-controller
      operationId: crear_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PedidoRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PedidoResponseDTO"
  /api/v1/categorias:
    get:
      tags:
      - categoria-controller
      operationId: listar_3
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCategoriaResponseDTO"
    post:
      tags:
      - categoria-controller
      operationId: crear_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoriaRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoriaResponseDTO"
  /api/v1/auth/register:
    post:
      tags:
      - auth-controller
      operationId: registrar
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponseDTO"
  /api/v1/auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequestDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponseDTO"
  /api/v1/pedidos/{id}/estado:
    patch:
      tags:
      - pedido-controller
      summary: Cambiar el estado de un pedido
      description: "Las unicas transiciones validas son: PENDIENTE -> CONFIRMADO o\
        \ CANCELADO; CONFIRMADO -> ENVIADO o CANCELADO; ENVIADO -> ENTREGADO. Cancelar\
        \ un pedido repone automaticamente el stock de sus lineas."
      operationId: cambiarEstado
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CambiarEstadoDTO"
        required: true
      responses:
        "404":
          description: Pedido no encontrado
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PedidoResponseDTO"
        "409":
          description: La transicion de estado solicitada no es valida
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PedidoResponseDTO"
        "200":
          description: Estado actualizado correctamente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PedidoResponseDTO"
  /api/v1/pedidos/{id}:
    get:
      tags:
      - pedido-controller
      operationId: obtenerPorId_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PedidoResponseDTO"
    delete:
      tags:
      - pedido-controller
      operationId: eliminar_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/categorias/{id}:
    get:
      tags:
      - categoria-controller
      operationId: obtenerPorId_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CategoriaResponseDTO"
    delete:
      tags:
      - categoria-controller
      operationId: eliminar_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
components:
  schemas:
    UsuarioRequestDTO:
      required:
      - email
      - nombre
      - password
      - rol
      type: object
      properties:
        nombre:
          maxLength: 150
          minLength: 0
          type: string
        email:
          type: string
        password:
          maxLength: 2147483647
          minLength: 8
          type: string
        rol:
          type: string
          enum:
          - ADMIN
          - VENDEDOR
          - CLIENTE
    UsuarioResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        email:
          type: string
        rol:
          type: string
          enum:
          - ADMIN
          - VENDEDOR
          - CLIENTE
    ProductoRequestDTO:
      required:
      - categoriaId
      - nombre
      - precio
      - stock
      type: object
      properties:
        nombre:
          maxLength: 150
          minLength: 0
          type: string
          description: Nombre del producto
          example: Teclado mecanico
        precio:
          type: number
          description: Precio de venta en euros
          example: 59.99
        stock:
          type: integer
          description: Unidades disponibles en stock
          format: int32
          example: 25
        categoriaId:
          type: integer
          description: Id de la categoria a la que pertenece el producto
          format: int64
          example: 1
    ProductoResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        precio:
          type: number
        stock:
          type: integer
          format: int32
        categoriaNombre:
          type: string
    LineaPedidoRequestDTO:
      required:
      - cantidad
      - productoId
      type: object
      properties:
        productoId:
          type: integer
          format: int64
        cantidad:
          type: integer
          format: int32
    PedidoRequestDTO:
      required:
      - lineas
      type: object
      properties:
        usuarioId:
          type: integer
          format: int64
        lineas:
          type: array
          items:
            $ref: "#/components/schemas/LineaPedidoRequestDTO"
    LineaPedidoResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productoNombre:
          type: string
        cantidad:
          type: integer
          format: int32
        precioUnitario:
          type: number
        subtotal:
          type: number
    PedidoResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        usuarioNombre:
          type: string
        fecha:
          type: string
          format: date-time
        estado:
          type: string
          enum:
          - PENDIENTE
          - CONFIRMADO
          - ENVIADO
          - ENTREGADO
          - CANCELADO
        total:
          type: number
        lineas:
          type: array
          items:
            $ref: "#/components/schemas/LineaPedidoResponseDTO"
    CategoriaRequestDTO:
      required:
      - nombre
      type: object
      properties:
        nombre:
          maxLength: 100
          minLength: 0
          type: string
    CategoriaResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
    RegisterRequestDTO:
      required:
      - email
      - nombre
      - password
      type: object
      properties:
        nombre:
          maxLength: 150
          minLength: 0
          type: string
        email:
          type: string
        password:
          maxLength: 2147483647
          minLength: 8
          type: string
    AuthResponseDTO:
      type: object
      properties:
        token:
          type: string
        tipo:
          type: string
    LoginRequestDTO:
      required:
      - email
      - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    CambiarEstadoDTO:
      required:
      - estado
      type: object
      properties:
        estado:
          type: string
          enum:
          - PENDIENTE
          - CONFIRMADO
          - ENVIADO
          - ENTREGADO
          - CANCELADO
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageUsuarioResponseDTO:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/UsuarioResponseDTO"
        number:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        sort:
          $ref: "#/components/schemas/SortObject"
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        sort:
          $ref: "#/components/schemas/SortObject"
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    PageProductoResponseDTO:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/ProductoResponseDTO"
        number:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        sort:
          $ref: "#/components/schemas/SortObject"
        empty:
          type: boolean
    PagePedidoResponseDTO:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/PedidoResponseDTO"
        number:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        sort:
          $ref: "#/components/schemas/SortObject"
        empty:
          type: boolean
    PageCategoriaResponseDTO:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/CategoriaResponseDTO"
        number:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        sort:
          $ref: "#/components/schemas/SortObject"
        empty:
          type: boolean
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT